+++ src/minifat.c
@@ -346,8 +346,8 @@ unsigned int fatOpenFile(FAT *fat, const char *fnm, FATFILE *fatfile)
  * fatReadFile - read from an already-open file.
  */
 
-int fatReadFile(FAT *fat, FATFILE *fatfile, void *buf, size_t szbuf,
-                size_t *readbytes)
+int fatReadFile(FAT *fat, FATFILE *fatfile, void *buf, unsigned int szbuf,
+                unsigned int *readbytes)
 {
     size_t bytesRead = 0;
     static unsigned char bbuf[MAXSECTSZ];
